requirement.txt
apache-airflow==2.9.2
requests
pandas
sqlalchemy
psycopg2-binary